+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_training_async.py
@@ -43,7 +43,7 @@ class TestTrainingAsync(AsyncFormRecognizerTest):
             self.assertIsNotNone(doc.page_count)
             self.assertEqual(doc.status, "succeeded")
             self.assertEqual(doc.errors, [])
-        for sub in model.models:
+        for sub in model.submodels:
             self.assertIsNotNone(sub.form_type)
             for key, field in sub.fields.items():
                 self.assertIsNotNone(field.label)
@@ -65,7 +65,7 @@ class TestTrainingAsync(AsyncFormRecognizerTest):
             self.assertIsNotNone(doc.page_count)
             self.assertEqual(doc.status, "succeeded")
             self.assertEqual(doc.errors, [])
-        for sub in model.models:
+        for sub in model.submodels:
             self.assertIsNotNone(sub.form_type)
             for key, field in sub.fields.items():
                 self.assertIsNotNone(field.label)
@@ -126,7 +126,7 @@ class TestTrainingAsync(AsyncFormRecognizerTest):
             self.assertIsNotNone(doc.page_count)
             self.assertEqual(doc.status, "succeeded")
             self.assertEqual(doc.errors, [])
-        for sub in model.models:
+        for sub in model.submodels:
             self.assertIsNotNone(sub.form_type)
             for key, field in sub.fields.items():
                 self.assertIsNotNone(field.accuracy)
@@ -148,7 +148,7 @@ class TestTrainingAsync(AsyncFormRecognizerTest):
             self.assertIsNotNone(doc.page_count)
             self.assertEqual(doc.status, "succeeded")
             self.assertEqual(doc.errors, [])
-        for sub in model.models:
+        for sub in model.submodels:
             self.assertIsNotNone(sub.form_type)
             self.assertIsNotNone(sub.accuracy)
             for key, field in sub.fields.items():