+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -17,6 +17,7 @@
 - `use_training_labels` is now a required positional param in the `begin_training` APIs.
 - `stream` and `url` parameters found on methods for `FormRecognizerClient` have been renamed to `form` and `form_url`, respectively.
 For recognize receipt methods, parameters have been renamed to `receipt` and `receipt_url`.
+- `models` property of `CustomFormModel` is renamed to `submodels`
 
 **New features**